api/requirements.txt
pydantic==2.9.1
openai==1.45.0
python-dotenv==1.0.1
fastapi[standard]==0.114.2
